requirements.txt
python-telegram-bot==13.15
feedparser==6.0.11
python-dotenv==1.0.1
deep-translator==1.11.4